+++ runtime/common/RuntimeMLIR.h
@@ -8,19 +8,16 @@
 
 #pragma once
 
-#include "JIT.h"
 #include "mlir/Tools/mlir-translate/Translation.h"
+#include <functional>
 #include <memory>
+#include <optional>
+#include <string>
 
 namespace mlir {
 class MLIRContext;
-class ExecutionEngine;
 class ModuleOp;
 } // namespace mlir
-
-namespace llvm {
-class Module;
-}
 
 namespace cudaq {
 
@@ -48,20 +45,6 @@ mlir::MLIRContext *getMLIRContext();
 /// @brief Create a new context and transfer the ownership. To be used to avoid
 /// overcrowding the current MLIR context with temporary modules.
 std::unique_ptr<mlir::MLIRContext> getOwningMLIRContext();
-
-/// @brief Given an LLVM Module, set its target triple corresponding to the
-/// current host machine.
-bool setupTargetTriple(llvm::Module *);
-
-/// @brief Run the LLVM PassManager.
-void optimizeLLVM(llvm::Module *);
-
-/// @brief Lower ModuleOp to a full QIR LLVMIR representation
-/// and return a module for JIT function pointer
-/// execution. JitEngine is automatically deleted when its resource count
-/// goes to zero.
-JitEngine createQIRJITEngine(mlir::ModuleOp &moduleOp,
-                             llvm::StringRef convertTo);
 
 class Translation {
 public:
